backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-generativeai
langchain
langchain-openai
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
tiktoken
pydantic
aiofiles
slowapi
cryptography
python-jose[cryptography]
passlib[bcrypt]